+++ lib/enemy/simple/simple_enemy.dart
@@ -1,43 +1,13 @@
-import 'package:bonfire/bonfire.dart';
-import 'package:flame/animation.dart' as FlameAnimation;
+import 'package:bonfire/enemy/enemy.dart';
+import 'package:bonfire/util/collision/collision.dart';
+import 'package:bonfire/util/direction.dart';
+import 'package:bonfire/util/direction_animations/simple_animation_enum.dart';
+import 'package:bonfire/util/direction_animations/simple_direction_animation.dart';
+import 'package:flame/position.dart';
 import 'package:flutter/widgets.dart';
 
 class SimpleEnemy extends Enemy {
-  /// Animation that was used when enemy stay stopped on the right.
-  final FlameAnimation.Animation animIdleRight;
-
-  /// Animation that was used when enemy stay stopped on the left.
-  final FlameAnimation.Animation animIdleLeft;
-
-  /// Animation that was used when enemy stay stopped on the top.
-  final FlameAnimation.Animation animIdleTop;
-
-  /// Animation that was used when enemy stay stopped on the bottom.
-  final FlameAnimation.Animation animIdleBottom;
-
-  /// Animation used when the enemy walks to the top.
-  final FlameAnimation.Animation animRunTop;
-
-  /// Animation used when the enemy walks to the right.
-  final FlameAnimation.Animation animRunRight;
-
-  /// Animation used when the enemy walks to the left.
-  final FlameAnimation.Animation animRunLeft;
-
-  /// Animation used when the enemy walks to the bottom.
-  final FlameAnimation.Animation animRunBottom;
-
-  final FlameAnimation.Animation animRunTopLeft;
-  final FlameAnimation.Animation animRunBottomLeft;
-
-  final FlameAnimation.Animation animRunTopRight;
-  final FlameAnimation.Animation animRunBottomRight;
-
-  final FlameAnimation.Animation animIdleTopLeft;
-  final FlameAnimation.Animation animIdleBottomLeft;
-
-  final FlameAnimation.Animation animIdleTopRight;
-  final FlameAnimation.Animation animIdleBottomRight;
+  SimpleDirectionAnimation animation;
 
   /// Variable that represents the speed of the enemy.
   final double speed;
@@ -56,22 +26,7 @@ class SimpleEnemy extends Enemy {
     @required Position initPosition,
     @required double height,
     @required double width,
-    @required this.animIdleRight,
-    @required this.animIdleLeft,
-    @required this.animRunRight,
-    @required this.animRunLeft,
-    this.animRunTopLeft,
-    this.animRunBottomLeft,
-    this.animRunTopRight,
-    this.animRunBottomRight,
-    this.animIdleTop,
-    this.animIdleBottom,
-    this.animRunTop,
-    this.animRunBottom,
-    this.animIdleTopLeft,
-    this.animIdleBottomLeft,
-    this.animIdleTopRight,
-    this.animIdleBottomRight,
+    @required this.animation,
     double life = 100,
     this.speed = 100,
     Collision collision,
@@ -94,10 +49,13 @@ class SimpleEnemy extends Enemy {
 
     if ((lastDirection != Direction.top || _isIdle) && addAnimation) {
       _isIdle = false;
-      animation = animRunTop ??
-          (lastDirectionHorizontal == Direction.right
-              ? animRunRight
-              : animRunLeft);
+      if (animation?.runTop != null) {
+        animation?.play(SimpleAnimationEnum.runTop);
+      } else {
+        animation?.play(lastDirectionHorizontal == Direction.right
+            ? SimpleAnimationEnum.runRight
+            : SimpleAnimationEnum.runLeft);
+      }
       lastDirection = Direction.top;
     }
   }
@@ -107,10 +65,13 @@ class SimpleEnemy extends Enemy {
     this.moveBottom(speed);
     if ((lastDirection != Direction.bottom || _isIdle) && addAnimation) {
       _isIdle = false;
-      animation = animRunBottom ??
-          (lastDirectionHorizontal == Direction.right
-              ? animRunRight
-              : animRunLeft);
+      if (animation?.runBottom != null) {
+        animation?.play(SimpleAnimationEnum.runBottom);
+      } else {
+        animation?.play(lastDirectionHorizontal == Direction.right
+            ? SimpleAnimationEnum.runRight
+            : SimpleAnimationEnum.runLeft);
+      }
       lastDirection = Direction.bottom;
     }
   }
@@ -120,7 +81,7 @@ class SimpleEnemy extends Enemy {
     this.moveLeft(speed);
     if ((lastDirection != Direction.left || _isIdle) && addAnimation) {
       _isIdle = false;
-      animation = animRunLeft;
+      animation?.play(SimpleAnimationEnum.runLeft);
       lastDirection = Direction.left;
       lastDirectionHorizontal = Direction.left;
     }
@@ -131,7 +92,7 @@ class SimpleEnemy extends Enemy {
     this.moveRight(speed);
     if ((lastDirection != Direction.right || _isIdle) && addAnimation) {
       _isIdle = false;
-      animation = animRunRight;
+      animation?.play(SimpleAnimationEnum.runRight);
       lastDirection = Direction.right;
       lastDirectionHorizontal = Direction.right;
     }
@@ -139,41 +100,34 @@ class SimpleEnemy extends Enemy {
 
   void customMoveTopRight(double speedX, double speedY) {
     if (_runFastAnimation) return;
-    if (animRunTopRight != null) {
-      animation = animRunTopRight;
-    }
+    animation?.play(SimpleAnimationEnum.runTopRight);
     lastDirection = Direction.topRight;
-    this.customMoveRight(speedX, addAnimation: animRunTopRight == null);
+    this.customMoveRight(speedX, addAnimation: animation?.runTopRight == null);
     this.customMoveTop(speedY, addAnimation: false);
   }
 
   void customMoveTopLeft(double speedX, double speedY) {
     if (_runFastAnimation) return;
-    if (animRunTopLeft != null) {
-      animation = animRunTopLeft;
-    }
+    animation?.play(SimpleAnimationEnum.runTopLeft);
     lastDirection = Direction.topLeft;
-    this.customMoveLeft(speedX, addAnimation: animRunTopLeft == null);
+    this.customMoveLeft(speedX, addAnimation: animation?.runTopLeft == null);
     this.customMoveTop(speedY, addAnimation: false);
   }
 
   void customMoveBottomRight(double speedX, double speedY) {
     if (_runFastAnimation) return;
-    if (animRunBottomRight != null) {
-      animation = animRunBottomRight;
-    }
+    animation?.play(SimpleAnimationEnum.runBottomRight);
     lastDirection = Direction.bottomRight;
-    this.customMoveRight(speedX, addAnimation: animRunBottomRight == null);
+    this.customMoveRight(speedX,
+        addAnimation: animation?.runBottomRight == null);
     this.customMoveBottom(speedY, addAnimation: false);
   }
 
   void customMoveBottomLeft(double speedX, double speedY) {
     if (_runFastAnimation) return;
-    if (animRunBottomLeft != null) {
-      animation = animRunBottomLeft;
-    }
+    animation?.play(SimpleAnimationEnum.runBottomLeft);
     lastDirection = Direction.bottomLeft;
-    this.customMoveLeft(speedX, addAnimation: animRunBottomLeft == null);
+    this.customMoveLeft(speedX, addAnimation: animation?.runBottomLeft == null);
     this.customMoveBottom(speedY, addAnimation: false);
   }
 
@@ -182,77 +136,77 @@ class SimpleEnemy extends Enemy {
     _isIdle = true;
     switch (lastDirection) {
       case Direction.left:
-        animation = animIdleLeft;
+        animation?.play(SimpleAnimationEnum.idleLeft);
         break;
       case Direction.right:
-        animation = animIdleRight;
+        animation?.play(SimpleAnimationEnum.idleRight);
         break;
       case Direction.top:
-        if (animIdleTop != null) {
-          animation = animIdleTop;
+        if (animation?.idleTop != null) {
+          animation?.play(SimpleAnimationEnum.idleTop);
         } else {
           if (lastDirectionHorizontal == Direction.left) {
-            animation = animIdleLeft;
+            animation?.play(SimpleAnimationEnum.idleLeft);
           } else {
-            animation = animIdleRight;
+            animation?.play(SimpleAnimationEnum.idleRight);
           }
         }
         break;
       case Direction.bottom:
-        if (animIdleBottom != null) {
-          animation = animIdleBottom;
+        if (animation?.idleBottom != null) {
+          animation?.play(SimpleAnimationEnum.idleBottom);
         } else {
           if (lastDirectionHorizontal == Direction.left) {
-            animation = animIdleLeft;
+            animation?.play(SimpleAnimationEnum.idleLeft);
           } else {
-            animation = animIdleRight;
+            animation?.play(SimpleAnimationEnum.idleRight);
           }
         }
         break;
       case Direction.topLeft:
-        if (animIdleTopLeft != null) {
-          animation = animIdleTopLeft;
+        if (animation?.idleTopLeft != null) {
+          animation?.play(SimpleAnimationEnum.idleTopLeft);
         } else {
-          if (animIdleLeft != null) animation = animIdleLeft;
+          animation?.play(SimpleAnimationEnum.idleLeft);
         }
         break;
       case Direction.topRight:
-        if (animIdleTopRight != null) {
-          animation = animIdleTopRight;
+        if (animation?.idleTopRight != null) {
+          animation?.play(SimpleAnimationEnum.idleTopRight);
         } else {
-          if (animIdleRight != null) animation = animIdleRight;
+          animation?.play(SimpleAnimationEnum.idleRight);
         }
         break;
       case Direction.bottomLeft:
-        if (animIdleBottomLeft != null) {
-          animation = animIdleBottomLeft;
+        if (animation?.idleBottomLeft != null) {
+          animation?.play(SimpleAnimationEnum.idleBottomLeft);
         } else {
-          if (animIdleLeft != null) animation = animIdleLeft;
+          animation?.play(SimpleAnimationEnum.idleLeft);
         }
         break;
       case Direction.bottomRight:
-        if (animIdleBottomRight != null) {
-          animation = animIdleBottomRight;
+        if (animation?.idleBottomRight != null) {
+          animation?.play(SimpleAnimationEnum.idleBottomRight);
         } else {
-          if (animIdleRight != null) animation = animIdleRight;
+          animation?.play(SimpleAnimationEnum.idleRight);
         }
         break;
     }
   }
 
-  void addFastAnimation(FlameAnimation.Animation animation,
-      {VoidCallback onFinish}) {
-    _runFastAnimation = true;
-    AnimatedObjectOnce fastAnimation = AnimatedObjectOnce(
-      animation: animation,
-      onlyUpdate: true,
-      onFinish: () {
-        _runFastAnimation = false;
-        idle();
-        if (onFinish != null) onFinish();
-      },
-    );
-    this.animation = fastAnimation.animation;
-    gameRef.add(fastAnimation);
+  @override
+  void update(double dt) {
+    if (isVisibleInCamera()) {
+      animation?.update(dt);
+    }
+    super.update(dt);
+  }
+
+  @override
+  void render(Canvas canvas) {
+    if (isVisibleInCamera()) {
+      animation?.render(canvas, position);
+    }
+    super.render(canvas);
   }
 }