+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.7.0]
+* BREAKING CHANGE: improvement in animations to SimplePlayer and SimpleEnemy. Now use `SimpleDirectionAnimation` to manipulate animations.
+* add `GameColorFilter`. It is now possible to add color filter in the game.
+* Possible to load maps made by Tiled from url. Just pass the link as path.
+
 ## [0.6.27]
 * little performance improvement;
 * remove mandatory joystick in widget;